requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
ecdsa==0.19.1
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
pillow==11.2.1
qrcode==8.1
six==1.17.0
Werkzeug==3.1.3
